% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeB}
\alias{computeB}
\title{Compute Sparse Matrix B}
\usage{
computeB(s, D_list, Dt, Ds, f, df, h)
}
\arguments{
\item{s}{A numeric matrix (\code{Eigen::MatrixXd}) representing grid points.
Each row corresponds to a grid point, and each column corresponds to a dimension.}

\item{D_list}{A list of sparse matrices (\code{Eigen::SparseMatrix}) representing
differential operators for each dimension.}

\item{Dt}{A numeric value specifying the time step parameter.}

\item{Ds}{A numeric value specifying the grid spacing parameter.}

\item{f}{A user-defined function applied to each row of \eqn{s}. It should return
a numeric vector.}

\item{df}{A user-defined derivative function applied to each row of \eqn{s}.
It should return a numeric vector.}

\item{h}{A user-defined function applied to each row of \eqn{s}. It contributes
to the diagonal updates in matrix \eqn{Q}. It should return a numeric vector.}
}
\value{
A sparse matrix (\code{Eigen::SparseMatrix}) representing the computed \eqn{B} matrix.
}
\description{
This function computes a sparse matrix \eqn{B} using input matrices, user-defined functions,
and sparse Kronecker-product-related operations. The matrix \eqn{B} is constructed by applying
transformations to input grid points and updating a diagonal matrix \eqn{Q} based on user-defined
functions and sparse matrices.
}
\details{
The function calculates the matrix \eqn{B} as follows:
\enumerate{
\item Apply the user-defined function \eqn{f} to each row of \eqn{s} to compute
the negated values stored in \eqn{p}.
\item Construct a diagonal matrix \eqn{Q}, where each diagonal element is computed
as the sum of the user-defined derivative function \eqn{df} and the inner product
of the function \eqn{h} applied to the corresponding row of \eqn{s}.
\item Update \eqn{Q} by adding contributions from \eqn{p} and the input list of
sparse matrices \eqn{D_list}.
\item Compute \eqn{B} as \eqn{I + Dt \cdot Q}, where \eqn{I} is an identity matrix.
}
}
