% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simulate_State_Obser}
\alias{Simulate_State_Obser}
\title{Simulate State and Observations}
\usage{
Simulate_State_Obser(Dt, Ntau, NtNtau, f, h, Dim, seed = NULL)
}
\arguments{
\item{Dt}{Numeric. The time step size for the simulation.}

\item{Ntau}{Integer. The number of observation points.}

\item{NtNtau}{Integer. The total number of time steps in the simulation.}

\item{f}{Function. A user-provided function representing the state dynamics.}

\item{h}{Function. A user-provided function representing the observation model.}

\item{Dim}{Integer. The dimension of the system.}

\item{seed}{Nullable Integer. An optional random seed for reproducibility.}
}
\value{
A list containing:
\item{x}{A matrix of the state trajectory over time.}
\item{y}{A matrix of the observations at specific time points.}
}
\description{
This function simulates the evolution of a system over time using the provided dynamics and observation models.
}
\details{
Simulate State and Observations
}
