% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_wrapper}
\alias{simulate_wrapper}
\title{Run the state simulation from C++ side}
\usage{
simulate_wrapper(Dt, Ntau, NtNtau, f, h, Dim, seed = NULL)
}
\arguments{
\item{Dt}{numeric, time step}

\item{Ntau}{integer, number of observation points}

\item{NtNtau}{integer, total steps}

\item{f}{function, dynamics}

\item{h}{function, observation}

\item{Dim}{integer, dimension}

\item{seed}{integer, random seed}
}
\value{
list with elements \code{x}, \code{y}, and \code{seed}
}
\description{
Run the state simulation from C++ side
}
